Harden core edges
- The no-cancel fast path now writes git's stdin on its own thread like
  the guarded path, so a large patch can't deadlock against a full
  stdout pipe.
- reset_worktree's throwaway index gets the pid+counter suffix the
  sequence-editor todos already used (shared unique_temp_suffix), so
  concurrent resets can't clobber each other.
- The sequence-editor path is shell-quoted instead of trusting temp_dir
  to contain no quotes.
- Bare 'git remote' is classified as a query, so the pickers' remote
  listing stops showing in the $ command log.
- FileEntry::is_unstaged -> has_worktree_changes: it includes untracked
  and conflicted entries, which the Unstaged *section* excludes — the
  old name invited that mixup.
- HeadInfo ahead/behind counts are u32 (they're never negative), and
  git_common_dir returns Result like git_dir instead of swallowing
  errors as None.

+++ crates/magritte/src/render.rs
@@ -887,8 +887,8 @@ impl StatusView {
         key: &str,
         glyph: &str,
         name: &str,
-        ahead: i64,
-        behind: i64,
+        ahead: u32,
+        behind: u32,
     ) -> gpui::Div {
         let mut chunk = div()
             .flex()
